cleaning up controller
