+++ go/controllers/auth.go
@@ -84,9 +84,9 @@ func (cfg *Config) Login(w http.ResponseWriter, r *http.Request) {
 	}
 
 	ctx := context.Background()
-	querier := db.New(cfg.DB)
+	query := db.New(cfg.DB)
 
-	user, err := querier.GetUserByEmail(ctx, req.Email)
+	user, err := query.GetUserByEmail(ctx, req.Email)
 	if err != nil {
 		http.Error(w, "Invalid email or password", http.StatusUnauthorized)
 		return
@@ -149,10 +149,10 @@ func (cfg *Config) Signup(w http.ResponseWriter, r *http.Request) {
 	}
 
 	ctx := context.Background()
-	querier := db.New(cfg.DB)
+	query := db.New(cfg.DB)
 
 	// Check if username already exists
-	_, err = querier.GetUserByUsername(ctx, req.Username)
+	_, err = query.GetUserByUsername(ctx, req.Username)
 	if err == nil {
 		http.Error(w, "Username already exists", http.StatusConflict)
 		return
@@ -163,7 +163,7 @@ func (cfg *Config) Signup(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Check if email already exists
-	_, err = querier.GetUserByEmail(ctx, req.Email)
+	_, err = query.GetUserByEmail(ctx, req.Email)
 	if err == nil {
 		http.Error(w, "Email already exists", http.StatusConflict)
 		return
@@ -181,7 +181,7 @@ func (cfg *Config) Signup(w http.ResponseWriter, r *http.Request) {
 		LastName:  req.LastName,
 	}
 
-	user, err := querier.CreateUser(ctx, params)
+	user, err := query.CreateUser(ctx, params)
 	if err != nil {
 		log.Printf("Error creating user: %v", err)
 		http.Error(w, "Error creating user", http.StatusInternalServerError)
@@ -217,146 +217,6 @@ func (cfg *Config) Signup(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusCreated)
 	json.NewEncoder(w).Encode(response)
-}
-
-// GetUser handles retrieving user information
-func (cfg *Config) GetUser(w http.ResponseWriter, r *http.Request) {
-	// Get user_id from context (set by AuthMiddleware)
-	userID, ok := r.Context().Value("user_id").(int32)
-	if !ok {
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	ctx := context.Background()
-	querier := db.New(cfg.DB)
-
-	user, err := querier.GetUserByID(ctx, userID)
-	if err != nil {
-		log.Printf("Error getting user: %v", err)
-		http.Error(w, "User not found", http.StatusNotFound)
-		return
-	}
-
-	// Convert to response type to avoid sending sensitive information
-	response := User{
-		ID:        user.ID,
-		Username:  user.Username,
-		Email:     user.Email,
-		FirstName: user.FirstName,
-		LastName:  user.LastName,
-		CreatedAt: user.CreatedAt.Time,
-		UpdatedAt: user.UpdatedAt.Time,
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(response)
-}
-
-// UpdateUser handles updating user information
-func (cfg *Config) UpdateUser(w http.ResponseWriter, r *http.Request) {
-	// Get user_id from context (set by AuthMiddleware)
-	userID, ok := r.Context().Value("user_id").(int32)
-	if !ok {
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	var req UpdateUserRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		http.Error(w, "Invalid request body", http.StatusBadRequest)
-		return
-	}
-
-	// Validate request
-	if req.Username == "" || req.Email == "" {
-		http.Error(w, "Username and email are required", http.StatusBadRequest)
-		return
-	}
-
-	ctx := context.Background()
-	conn, err := cfg.DB.Acquire(ctx)
-	if err != nil {
-		log.Printf("could not connect to db... %v", err)
-		http.Error(w, "Database connection error", http.StatusInternalServerError)
-		return
-	}
-	defer conn.Release()
-
-	querier := db.New(conn)
-
-	// Check if username is taken (by another user)
-	existingUser, err := querier.GetUserByUsername(ctx, req.Username)
-	if err == nil && existingUser.ID != userID {
-		http.Error(w, "Username already taken", http.StatusConflict)
-		return
-	}
-
-	// Check if email is taken (by another user)
-	existingUser, err = querier.GetUserByEmail(ctx, req.Email)
-	if err == nil && existingUser.ID != userID {
-		http.Error(w, "Email already taken", http.StatusConflict)
-		return
-	}
-
-	// Update user
-	err = querier.UpdateUser(ctx, db.UpdateUserParams{
-		Username:  req.Username,
-		Email:     req.Email,
-		FirstName: req.FirstName,
-		LastName:  req.LastName,
-		ID:        userID,
-	})
-	if err != nil {
-		log.Printf("error updating user: %v", err)
-		http.Error(w, "Failed to update user", http.StatusInternalServerError)
-		return
-	}
-
-	// Get updated user info
-	user, err := querier.GetUserByID(ctx, userID)
-	if err != nil {
-		log.Printf("error getting updated user: %v", err)
-		http.Error(w, "Failed to get updated user info", http.StatusInternalServerError)
-		return
-	}
-
-	// Convert to response type
-	response := User{
-		ID:        user.ID,
-		Username:  user.Username,
-		Email:     user.Email,
-		FirstName: user.FirstName,
-		LastName:  user.LastName,
-		CreatedAt: user.CreatedAt.Time,
-		UpdatedAt: user.UpdatedAt.Time,
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(response)
-}
-
-// DeleteUser handles user account deletion
-func (cfg *Config) DeleteUser(w http.ResponseWriter, r *http.Request) {
-	// Get user_id from context (set by AuthMiddleware)
-	userID, ok := r.Context().Value("user_id").(int32)
-	if !ok {
-		http.Error(w, "Unauthorized", http.StatusUnauthorized)
-		return
-	}
-
-	ctx := context.Background()
-	querier := db.New(cfg.DB)
-
-	err := querier.DeleteUser(ctx, userID)
-	if err != nil {
-		log.Printf("Error deleting user: %v", err)
-		http.Error(w, "Error deleting user", http.StatusInternalServerError)
-		return
-	}
-
-	// Return success with no content
-	w.WriteHeader(http.StatusNoContent)
 }
 
 // AuthMiddleware handles authentication for incoming requests
